Dividing gateway code into different units
