+++ gateways/custom/storagegrid/stop_test.go
@@ -1,0 +1,20 @@
+package storagegrid
+
+import (
+	"testing"
+	"github.com/argoproj/argo-events/gateways"
+	"github.com/stretchr/testify/assert"
+)
+
+func TestStorageGridConfigExecutor_StopConfig(t *testing.T) {
+	s3Config := &StorageGridConfigExecutor{}
+	ctx := &gateways.ConfigContext{}
+	ctx.StopChan = make(chan struct{})
+	ctx.Active = true
+	go func() {
+		msg := <-ctx.StopChan
+		assert.Equal(t, msg, struct{}{})
+	}()
+	s3Config.StopConfig(ctx)
+	assert.Equal(t, false, ctx.Active)
+}
