+++ gateways/errors.go
@@ -1,6 +1,8 @@
 package gateways
 
-import "errors"
+import (
+	"errors"
+)
 
 const (
 	ErrGatewayTransformerConnectionMsg = "failed to connect to gateway transformer"
@@ -10,5 +12,6 @@ const (
 
 var (
 	ErrConfigParseFailed = errors.New("failed to parse configuration")
+	ErrEmptyConfig = errors.New("configuration must be non empty")
 	ErrInvalidConfig     = errors.New("invalid configuration")
 )
